refactor(foldkit): split runtime helpers into focused modules
Move the helpers that never touch runtime closure state out of
runtime.ts into modules named for what they own, with each module's
types beside its code: the Dispatch service, visibility resolution,
DevTools config, slow-phase instrumentation, the duplicate id scanner,
crash rendering, document metadata, the hydration handoff, the host
connector, the HMR model bridge, and the browser scheduler.

RoutingConfig now lives with the navigation listeners, which removes
the import cycle between browserListeners.ts and runtime.ts. The
runtime barrel lists its exports explicitly and public.ts re-exports
through it, so the published surface is unchanged.

Beyond the moves:

- An isVisible helper replaces the two inline show checks with the
  same semantics.
- Four resolvers in devToolsConfig.ts wrap the DevTools option pipes
  that sat inline in makeRuntime, and the iframe check moves inside
  the config resolver.
- The default crash view takes the real CrashContext instead of a
  structural stand-in.
- hydrationHandoff.ts uses the DOM Document directly, since nothing
  there shadows it, so the DomDocument alias is gone.
- The scene test helpers import the Dispatch tag from dispatch.ts
  instead of loading the whole runtime graph.
