+++ packages/foldkit/src/devTools/host.ts
@@ -1,6 +1,6 @@
 export { DEVTOOLS_HOST_ID } from '../html/index.js'
-export { __setDevToolsOverlay } from '../runtime/runtime.js'
-export type { DevToolsOverlay } from '../runtime/runtime.js'
+export { __setDevToolsOverlay } from '../runtime/devToolsConfig.js'
+export type { DevToolsOverlay } from '../runtime/devToolsConfig.js'
 
 export { INIT_INDEX, latestEntryIndex } from './store.js'
