+++ packages/foldkit/src/runtime/browserListeners.ts
@@ -3,7 +3,12 @@ import { Option, String } from 'effect'
 import { OptionExt, StringExt } from '../effectExtensions/index.js'
 import { UrlRequest } from '../navigation/urlRequest.js'
 import { Url } from '../url/index.js'
-import { RoutingConfig } from './runtime.js'
+
+/** Configuration for URL routing with handlers for URL requests and URL changes. */
+export type RoutingConfig<Message> = Readonly<{
+  onUrlRequest: (request: UrlRequest) => Message
+  onUrlChange: (url: Url) => Message
+}>
 
 export const addNavigationEventListeners = <Message>(
   dispatch: (message: Message) => void,
