+++ packages/foldkit/src/runtime/dispatch.ts
@@ -1,0 +1,10 @@
+import { Context, Effect } from 'effect'
+
+/** Effect service tag that provides message dispatching to the view layer. */
+export class Dispatch extends Context.Service<
+  Dispatch,
+  {
+    readonly dispatchAsync: (message: unknown) => Effect.Effect<void>
+    readonly dispatchSync: (message: unknown) => void
+  }
+>()('@foldkit/Dispatch') {}
