+++ packages/foldkit/src/runtime/mountViewState.test.ts
@@ -27,11 +27,8 @@ import * as Mount from '../mount/index.js'
 import { evo } from '../struct/index.js'
 import * as Subscription from '../subscription/subscription.js'
 import type * as Update from '../update/index.js'
-import {
-  __setDevToolsOverlay,
-  makeApplication,
-  makeElement,
-} from './runtime.js'
+import { __setDevToolsOverlay } from './devToolsConfig.js'
+import { makeApplication, makeElement } from './runtime.js'
 
 const Message = defineMessageUnion({
   CompletedMountEditor: {},
