+++ packages/foldkit/src/runtime/runtime.ts
@@ -8,22 +8,18 @@ import {
   Fiber,
   Function,
   Layer,
-  Match,
   Option,
   Predicate,
   PubSub,
   Record,
   Ref,
   Runtime,
-  Scheduler,
   Schema,
-  Scope,
   Stream,
   SubscriptionRef,
   pipe,
 } from 'effect'
 
-import { HYDRATION_BUILD_ATTRIBUTE } from '../buildToken.js'
 import {
   __CurrentRegistry as __CurrentInterruptRegistry,
   __makeRegistry as __makeInterruptRegistry,
@@ -48,13 +44,9 @@ import {
   __flushReplayUnmountsAfterPatchFailure as flushReplayUnmountsAfterPatchFailure,
   __htmlBuilder as htmlBuilderFor,
   __setRuntime as setHtmlRuntime,
-  textDirectionToAttribute,
 } from '../html/index.js'
 import { __hydrateVNode } from '../hydrate.js'
-import {
-  FOLDKIT_APP_ATTRIBUTE,
-  FOLDKIT_FLAGS_ATTRIBUTE,
-} from '../hydrationMarker.js'
+import { FOLDKIT_APP_ATTRIBUTE } from '../hydrationMarker.js'
 import type {
   ManagedResourceConfig,
   ManagedResources,
@@ -65,16 +57,7 @@ import {
   ViewState,
   liveViewStateChanges,
 } from '../mount/index.js'
-import { UrlRequest } from '../navigation/urlRequest.js'
-import {
-  type Inbound,
-  type Outbound,
-  type Ports,
-  __CurrentPortChannels,
-  type __InboundChannel,
-  type __PortChannels,
-  __makeInboundChannel,
-} from '../port/index.js'
+import { type Ports, __CurrentPortChannels } from '../port/index.js'
 import { RenderCommit, createCommitNotifier } from '../render/commit.js'
 import type { Subscriptions } from '../subscription/subscription.js'
 import type { Return as UpdateReturn } from '../update/index.js'
@@ -84,19 +67,62 @@ import {
   __patchVNode,
   __recoverVNodeAfterPatchFailure,
 } from '../vdom.js'
-import { addNavigationEventListeners } from './browserListeners.js'
-import { defaultCrashView, noOpDispatch } from './crashUI.js'
+import {
+  type RoutingConfig,
+  addNavigationEventListeners,
+} from './browserListeners.js'
+import { provideBrowserScheduler } from './browserScheduler.js'
+import {
+  type CrashConfig,
+  type CrashContext,
+  type ElementCrashConfig,
+  type VNodeSlot,
+  noOpDispatch,
+  renderCrashView,
+} from './crashUI.js'
 import { deepFreeze } from './deepFreeze.js'
 import {
-  PreserveModelMessage,
-  RequestModelMessage,
-  RestoreModelMessage,
-} from './hmrProtocol.js'
+  type DevToolsConfig,
+  resolveDevToolsConfig,
+  resolveDevToolsKeyframeInterval,
+  resolveDevToolsMaxEntries,
+  resolveExcludeFromHistoryTags,
+} from './devToolsConfig.js'
+import { Dispatch } from './dispatch.js'
+import { applyDocumentMetadata } from './documentMetadata.js'
+import { createDuplicateIdScanner } from './duplicateIdScanner.js'
+import { preserveModel, resolveHmrModel } from './hmrModelBridge.js'
 import {
   preserveScrollPosition,
   restorePreservedScrollPosition,
 } from './hmrScroll.js'
+import {
+  type EmbedHandle,
+  type HostConnector,
+  type PortChannelsBundle,
+  buildPortHandles,
+  makeHostConnector,
+  makePortChannels,
+  validatePorts,
+} from './hostConnector.js'
+import {
+  type HydrationConfig,
+  buildSkew,
+  containRefusedPage,
+  findDocumentHydration,
+  hasServerRenderedMarkup,
+} from './hydrationHandoff.js'
 import { makePreserveScheduler } from './preserveScheduler.js'
+import {
+  type SlowConfig,
+  type SlowPatchContext,
+  type SlowSubscriptionDependenciesContext,
+  type SlowUpdateContext,
+  type SlowViewContext,
+  __resolveSlowConfig,
+  measureSlowPhase,
+  reportSlowPhase,
+} from './slowPhase.js'
 import {
   type StartViewTransition,
   type ViewTransitionConfig,
@@ -105,6 +131,7 @@ import {
   __resolveStartViewTransition,
   __silenceViewTransitionRejections,
 } from './viewTransition.js'
+import { type Visibility, isVisible } from './visibility.js'
 
 type AnyCommand<T, E = never, R = never> = {
   readonly name: string
@@ -129,869 +156,9 @@ const toCommandRecord = (
     ? { name: command.name, args: command.args }
     : { name: command.name }
 
-/** Position of the DevTools badge and panel on screen. */
-export type DevToolsPosition =
-  'BottomRight' | 'BottomLeft' | 'TopRight' | 'TopLeft'
-
-/** Controls when a feature is shown. */
-export type Visibility = 'Development' | 'Always'
-
-/** Controls DevTools interaction mode.
- *
- * - `'Inspect'`: Messages stream in and clicking a row shows its state snapshot without pausing the rendered view.
- * - `'TimeTravel'`: Clicking a row installs a paused historical view while the live application continues. Resume to patch the latest live view.
- */
-export type DevToolsMode = 'Inspect' | 'TimeTravel'
-
-/** Mode value for the DevTools panel. Either a single mode used in every
- *  environment, or an object selecting different modes for development and
- *  production. Use the object form to keep `'TimeTravel'` for local debugging
- *  while shipping the safer `'Inspect'` mode to users. `'TimeTravel'` in
- *  production pauses the user's rendered view when a history row is clicked. */
-export type DevToolsModeConfig =
-  | DevToolsMode
-  | Readonly<{ development: DevToolsMode; production: DevToolsMode }>
-
-/**
- * Factory that mounts the in-browser DevTools overlay against a recording
- * store. The runtime keeps the store and the WebSocket bridge (so external
- * tooling like the DevTools MCP server works without an overlay); the visual
- * overlay is injected so it can live in `@foldkit/devtools` and pull in
- * `@foldkit/ui` without coupling the core runtime to either.
- *
- * Foldkit's Vite plugin supplies the installed overlay automatically. It is
- * included in production when `@foldkit/devtools` is a regular dependency.
- */
-export type DevToolsOverlay = (
-  store: DevToolsStore,
-  position: DevToolsPosition,
-  mode: DevToolsMode,
-  maybeBanner: Option.Option<string>,
-) => Effect.Effect<void, never, Scope.Scope>
-
-/**
- * DevTools configuration.
- *
- * Pass `false` to disable DevTools entirely.
- *
- * - `show`: `'Development'` (default) enables in dev mode only, `'Always'` enables in all environments including production.
- * - `position`: Where the badge and panel appear. Defaults to `'BottomRight'`.
- * - `mode`: `'TimeTravel'` (default) enables full time-travel debugging by installing a paused historical view while the live application continues. `'Inspect'` allows browsing state snapshots without replacing the live view. Pass `{ development, production }` to use different modes per environment. Useful when DevTools is shown in production (`show: 'Always'`) and you want `'TimeTravel'` only in local development.
- * - `banner`: Optional text shown as a banner at the top of the panel.
- * - `excludeFromHistory`: Message `_tag` values whose dispatches should not be recorded in DevTools history. The Messages still drive `update` and the runtime as usual; they just don't appear in the history panel and don't pay the per-Message diff cost. Use for high-frequency Messages (animation frames, pointer moves, scroll events) that would flood history without adding insight.
- * - `maxEntries`: Maximum number of recorded Messages retained in history before the oldest is evicted. Defaults to 100. Clamped to the range 20-500: smaller values keep the panel snappy under high message rates, larger values give you more scroll-back. Each retained entry stores a full Model snapshot, so memory cost scales linearly with both `maxEntries` and your Model size.
- * - `keyframeInterval`: Number of recorded Messages between full Model snapshots. Defaults to 31. Time-travel to an index replays `update` forward from the nearest earlier keyframe, so this is a memory/time tradeoff: smaller values store more snapshots (more memory) but make each jump cheaper, down to `1` where every jump is a constant-time snapshot lookup with no replay. Reach for a denser interval when the app has a heavy `update` and time-travel jumps feel sluggish. Clamped to a minimum of 1. Forced to 1 automatically when `excludeFromHistory` is active, since excluded Messages are never replayed.
- */
-export type DevToolsConfig =
-  | false
-  | Readonly<{
-      show?: Visibility
-      position?: DevToolsPosition
-      mode?: DevToolsModeConfig
-      banner?: string
-      excludeFromHistory?: ReadonlyArray<string>
-      maxEntries?: number
-      keyframeInterval?: number
-      /**
-       * The application's `Message` Schema. When provided and the running app
-       * is connected to the Foldkit DevTools MCP server, AI agents can dispatch
-       * Messages into the runtime. The Schema decodes inbound dispatch payloads
-       * at the bridge boundary and returns a clean error on mismatch.
-       *
-       * Without this field, `RequestDispatchMessage` is rejected with an
-       * informative error.
-       */
-      Message?: Schema.Codec<any, any, unknown, unknown>
-    }>
-
-let registeredDevToolsOverlay: DevToolsOverlay | undefined
-
-/** Registers the overlay supplied by the Foldkit Vite plugin. */
-export const __setDevToolsOverlay = (
-  overlay: DevToolsOverlay | undefined,
-): void => {
-  registeredDevToolsOverlay = overlay
-}
-
-const DEFAULT_DEV_TOOLS_SHOW: Visibility = 'Development'
-const DEFAULT_DEV_TOOLS_POSITION: DevToolsPosition = 'BottomRight'
-const DEFAULT_DEV_TOOLS_MODE: DevToolsMode = 'TimeTravel'
-
-const resolveDevToolsMode = (config: DevToolsModeConfig): DevToolsMode => {
-  if (typeof config === 'string') {
-    return config
-  } else {
-    return import.meta.hot ? config.development : config.production
-  }
-}
-const DEV_TOOLS_MAX_ENTRIES_MIN = 20
-const DEV_TOOLS_MAX_ENTRIES_MAX = 500
-const DEV_TOOLS_KEYFRAME_INTERVAL_MIN = 1
-
-/** Context provided when view construction exceeds its configured time budget. */
-export type SlowViewContext<Model, Message> = Readonly<{
-  _tag: 'View'
-  model: Model
-  message: Option.Option<Message>
-  durationMs: number
-  thresholdMs: number
-}>
-
-/** Context provided when update exceeds its configured time budget. */
-export type SlowUpdateContext<Model, Message> = Readonly<{
-  _tag: 'Update'
-  previousModel: Model
-  nextModel: Model
-  message: Message
-  durationMs: number
-  thresholdMs: number
-}>
-
-/** Context provided when DOM patching exceeds its configured time budget. */
-export type SlowPatchContext<Model, Message> = Readonly<{
-  _tag: 'Patch'
-  model: Model
-  message: Option.Option<Message>
-  durationMs: number
-  thresholdMs: number
-}>
-
-/** Context provided when subscription dependency extraction exceeds its configured time budget. */
-export type SlowSubscriptionDependenciesContext<Model> = Readonly<{
-  _tag: 'SubscriptionDependencies'
-  subscriptionKey: string
-  model: Model
-  durationMs: number
-  thresholdMs: number
-}>
-
-/** Tagged union of every slow-phase context passed to `slow.onSlow`. */
-export type SlowContext<Model, Message> =
-  | SlowViewContext<Model, Message>
-  | SlowUpdateContext<Model, Message>
-  | SlowPatchContext<Model, Message>
-  | SlowSubscriptionDependenciesContext<Model>
-
-/** Phase names measured by the slow warning runtime option. */
-export const SlowPhase = Schema.Literals([
-  'Update',
-  'View',
-  'Patch',
-  'SubscriptionDependencies',
-])
-export type SlowPhase = typeof SlowPhase.Type
-
-/** Budget overrides for slow warning phases. Omitted fields use Foldkit defaults. */
-export type SlowThresholdOverrides = Readonly<{
-  Update?: number
-  View?: number
-  Patch?: number
-  SubscriptionDependencies?: number
-}>
-
-type ResolvedSlowPhaseConfig<Context> = Readonly<{
-  thresholdMs: number
-  onSlow: (context: Context) => void
-}>
-
-type ResolvedSlowConfig<Model, Message> = Readonly<{
-  view: Option.Option<ResolvedSlowPhaseConfig<SlowViewContext<Model, Message>>>
-  update: Option.Option<
-    ResolvedSlowPhaseConfig<SlowUpdateContext<Model, Message>>
-  >
-  patch: Option.Option<
-    ResolvedSlowPhaseConfig<SlowPatchContext<Model, Message>>
-  >
-  subscriptionDependencies: Option.Option<
-    ResolvedSlowPhaseConfig<SlowSubscriptionDependenciesContext<Model>>
-  >
-}>
-
-/**
- * Slow-phase warning configuration.
- *
- * By default, all phases are enabled in development with Foldkit's default
- * thresholds. Pass `false` to disable warnings entirely. Pass an object to
- * refine those defaults.
- *
- * - `show`: `'Development'` (default) enables warnings only when Vite HMR is active. `'Always'` enables them in every environment.
- * - `measuredPhases`: Phases to measure. Defaults to every slow warning phase.
- * - `thresholdOverrides`: Per-phase budget overrides. Omitted fields keep defaults; overrides for unmeasured phases are ignored.
- * - `onSlow`: Callback for every measured phase that exceeds its budget. Replaces Foldkit's default `console.warn`; Foldkit will not also warn for tags your callback ignores.
- */
-export type SlowConfig<Model, Message> =
-  | false
-  | Readonly<{
-      show?: Visibility
-      measuredPhases?: ReadonlyArray<SlowPhase>
-      thresholdOverrides?: SlowThresholdOverrides
-      onSlow?: (context: SlowContext<Model, Message>) => void
-    }>
-
-const DEFAULT_SLOW_SHOW: Visibility = 'Development'
-const DEFAULT_SLOW_VIEW_THRESHOLD_MS = 16
-const DEFAULT_SLOW_UPDATE_THRESHOLD_MS = 4
-const DEFAULT_SLOW_PATCH_THRESHOLD_MS = 8
-const DEFAULT_SLOW_SUBSCRIPTION_DEPENDENCIES_THRESHOLD_MS = 2
-
-const ALL_SLOW_PHASES: ReadonlyArray<SlowPhase> = [
-  'Update',
-  'View',
-  'Patch',
-  'SubscriptionDependencies',
-]
-
-const resolveSlowPhase = <Context>(
-  isMeasured: boolean,
-  thresholdMs: number,
-  onSlow: (context: Context) => void,
-): Option.Option<ResolvedSlowPhaseConfig<Context>> =>
-  Option.liftPredicate(
-    {
-      thresholdMs,
-      onSlow,
-    },
-    () => isMeasured,
-  )
-
-export const __resolveSlowConfig = <Model, Message>(
-  slow: SlowConfig<Model, Message> | undefined,
-  isSlowVisible: (show: Visibility) => boolean,
-): Option.Option<ResolvedSlowConfig<Model, Message>> => {
-  const maybeSlowConfig = Match.value(slow).pipe(
-    Match.withReturnType<
-      Option.Option<Exclude<SlowConfig<Model, Message>, false>>
-    >(),
-    Match.when(false, () => Option.none()),
-    Match.when(Predicate.isUndefined, () =>
-      Option.some<Exclude<SlowConfig<Model, Message>, false>>({}),
-    ),
-    Match.orElse(config => Option.some(config)),
-  )
-
-  return pipe(
-    maybeSlowConfig,
-    Option.filter(config => isSlowVisible(config.show ?? DEFAULT_SLOW_SHOW)),
-    Option.map(config => {
-      const onSlow = config.onSlow ?? defaultSlowCallback
-      const measuredPhases = config.measuredPhases ?? ALL_SLOW_PHASES
-      const isPhaseMeasured = (phase: SlowPhase): boolean =>
-        Array.contains(measuredPhases, phase)
-
-      return {
-        view: resolveSlowPhase(
-          isPhaseMeasured('View'),
-          config.thresholdOverrides?.View ?? DEFAULT_SLOW_VIEW_THRESHOLD_MS,
-          onSlow,
-        ),
-        update: resolveSlowPhase(
-          isPhaseMeasured('Update'),
-          config.thresholdOverrides?.Update ?? DEFAULT_SLOW_UPDATE_THRESHOLD_MS,
-          onSlow,
-        ),
-        patch: resolveSlowPhase(
-          isPhaseMeasured('Patch'),
-          config.thresholdOverrides?.Patch ?? DEFAULT_SLOW_PATCH_THRESHOLD_MS,
-          onSlow,
-        ),
-        subscriptionDependencies: resolveSlowPhase(
-          isPhaseMeasured('SubscriptionDependencies'),
-          config.thresholdOverrides?.SubscriptionDependencies ??
-            DEFAULT_SLOW_SUBSCRIPTION_DEPENDENCIES_THRESHOLD_MS,
-          onSlow,
-        ),
-      }
-    }),
-  )
-}
-
-const measureSlowPhase = <Context, Result>(
-  maybeConfig: Option.Option<ResolvedSlowPhaseConfig<Context>>,
-  run: () => Result,
-): readonly [Result, Option.Option<number>] => {
-  if (Option.isSome(maybeConfig)) {
-    const start = performance.now()
-    const result = run()
-
-    return [result, Option.some(performance.now() - start)]
-  } else {
-    return [run(), Option.none()]
-  }
-}
-
-const reportSlowPhase = <Context>(
-  maybeConfig: Option.Option<ResolvedSlowPhaseConfig<Context>>,
-  maybeDurationMs: Option.Option<number>,
-  makeContext: (durationMs: number, thresholdMs: number) => Context,
-): void => {
-  if (Option.isSome(maybeConfig)) {
-    const { thresholdMs, onSlow } = maybeConfig.value
-    const maybeExceededDuration = Option.filter(
-      maybeDurationMs,
-      durationMs => durationMs > thresholdMs,
-    )
-
-    if (Option.isSome(maybeExceededDuration)) {
-      onSlow(makeContext(maybeExceededDuration.value, thresholdMs))
-    }
-  }
-}
-
-const messageTag = (rawMessage: unknown): string =>
-  pipe(
-    rawMessage,
-    Option.liftPredicate(Predicate.isObject),
-    Option.flatMap(Record.get('_tag')),
-    Option.match({
-      onNone: () => 'unknown',
-      onSome: String,
-    }),
-  )
-
-const optionMessageTrigger = (maybeMessage: Option.Option<unknown>): string =>
-  Option.match(maybeMessage, {
-    onNone: () => 'init',
-    onSome: messageTag,
-  })
-
-const TUNING_HINT =
-  'Set slow.thresholdOverrides to change budgets or pass slow: false to disable warnings.'
-
-export const defaultSlowCallback = (
-  context: SlowContext<unknown, unknown>,
-): void => {
-  const { durationMs, thresholdMs: budget } = context
-  const duration = durationMs.toFixed(1)
-
-  const summary = Match.value(context).pipe(
-    Match.tagsExhaustive({
-      View: ({ message }) =>
-        `Slow view: ${duration}ms (budget: ${budget}ms), triggered by ${optionMessageTrigger(message)}. Keep render-only work in the view path and memoize expensive subtrees with createLazy or createKeyedLazy.`,
-      Update: ({ message }) =>
-        `Slow update: ${duration}ms (budget: ${budget}ms), triggered by ${messageTag(message)}. Inspect the triggering Message branch; move render-only derivations to memoized views and keep update focused on state transitions.`,
-      Patch: ({ message }) =>
-        `Slow patch: ${duration}ms (budget: ${budget}ms), triggered by ${optionMessageTrigger(message)}. Key mapped lists by stable ids, split large views, or memoize stable subtrees with createLazy.`,
-      SubscriptionDependencies: ({ subscriptionKey }) =>
-        `Slow subscription dependencies: ${duration}ms (budget: ${budget}ms) for subscription "${subscriptionKey}". Keep modelToDependencies a cheap projection from modeled fields; avoid scans, sorting, serialization, and large dependency objects.`,
-    }),
-  )
-
-  const maybeRawMessage: Option.Option<unknown> = Match.value(context).pipe(
-    Match.withReturnType<Option.Option<unknown>>(),
-    Match.tagsExhaustive({
-      Update: ({ message }) => Option.some<unknown>(message),
-      View: ({ message }) => message,
-      Patch: ({ message }) => message,
-      SubscriptionDependencies: () => Option.none(),
-    }),
-  )
-
-  console.warn(
-    `[foldkit] ${summary} ${TUNING_HINT}`,
-    context,
-    ...Option.toArray(maybeRawMessage),
-  )
-}
-
-/** Dev-only scan for duplicate DOM ids within the Foldkit-rendered root.
- *
- *  Duplicate ids are invalid HTML but browsers do not report them, so
- *  `getElementById` / `querySelector` silently resolve to the first match and
- *  focus or ARIA labelling can target the wrong element with no error. The
- *  scan is scoped to `root` (not the whole document) so unrelated page ids do
- *  not trigger it, warns rather than throws, and dedupes through `warnedIds`
- *  so the same collision is reported at most once instead of every render. */
-const warnDuplicateIds = (
-  root: Node | undefined,
-  warnedIds: Set<string>,
-): void => {
-  if (!(root instanceof Element)) {
-    return
-  }
-
-  const seenIds = new Set<string>()
-  const duplicateIds = new Set<string>()
-
-  const elementsWithId = Array.fromIterable(root.querySelectorAll('[id]'))
-  if (root.id !== '') {
-    elementsWithId.unshift(root)
-  }
-
-  for (const element of elementsWithId) {
-    const { id } = element
-    if (seenIds.has(id)) {
-      duplicateIds.add(id)
-    } else {
-      seenIds.add(id)
-    }
-  }
-
-  for (const id of duplicateIds) {
-    if (!warnedIds.has(id)) {
-      warnedIds.add(id)
-      console.warn(
-        `[foldkit] Duplicate DOM id "${id}" in the rendered tree. Ids must be ` +
-          'unique within the document; otherwise focus and ARIA labelling can ' +
-          'silently target the wrong element. Give each element a unique id.',
-      )
-    }
-  }
-}
-
-const DUPLICATE_ID_SCAN_INTERVAL_MS = 1000
-
-type DuplicateIdScanner = Readonly<{
-  schedule: (root: Node | undefined) => void
-  cancel: () => void
-}>
-
-/** Coalesces `warnDuplicateIds` scans so rapid successive renders trigger at
- *  most one full-tree scan per `DUPLICATE_ID_SCAN_INTERVAL_MS`, bounding the
- *  cost under high-frequency dev rendering such as animationFrame subscriptions
- *  or per-keystroke input. A real duplicate id persists across renders, so
- *  scanning the latest tree on a trailing timer never misses a genuine
- *  collision, and the retained `warnedIds` keeps each id to a single warning.
- *  Dev-only: the runtime both creates the scanner and calls `schedule` behind
- *  `import.meta.hot`, so the scanner and its DOM scan tree-shake out of
- *  production builds. */
-const createDuplicateIdScanner = (): DuplicateIdScanner => {
-  const warnedIds = new Set<string>()
-  let timeoutHandle: ReturnType<typeof setTimeout> | undefined
-  let latestRoot: Node | undefined
-
-  const schedule = (root: Node | undefined): void => {
-    latestRoot = root
-    if (timeoutHandle !== undefined) {
-      return
-    }
-    timeoutHandle = setTimeout(() => {
-      timeoutHandle = undefined
-      warnDuplicateIds(latestRoot, warnedIds)
-    }, DUPLICATE_ID_SCAN_INTERVAL_MS)
-  }
-
-  const cancel = (): void => {
-    if (timeoutHandle !== undefined) {
-      clearTimeout(timeoutHandle)
-      timeoutHandle = undefined
-    }
-    latestRoot = undefined
-  }
-
-  return { schedule, cancel }
-}
-
-/** Effect service tag that provides message dispatching to the view layer. */
-export class Dispatch extends Context.Service<
-  Dispatch,
-  {
-    readonly dispatchAsync: (message: unknown) => Effect.Effect<void>
-    readonly dispatchSync: (message: unknown) => void
-  }
->()('@foldkit/Dispatch') {}
-
 export type { Command } from '../command/index.js'
 
-/** Configuration for URL routing with handlers for URL requests and URL changes. */
-export type RoutingConfig<Message> = Readonly<{
-  onUrlRequest: (request: UrlRequest) => Message
-  onUrlChange: (url: Url) => Message
-}>
-
-/** Context provided to crash.view and crash.report when the runtime encounters
- *  an unrecoverable error. `message` is the Message being processed when the
- *  crash occurred, present as an `Option` because a crash during the initial
- *  render has no triggering Message. */
-export type CrashContext<Model, Message> = Readonly<{
-  error: Error
-  model: Model
-  message: Option.Option<Message>
-}>
-
-/** Configuration for crash handling, with custom crash UI and/or crash
- *  reporting. The crash view renders after the dispatch loop has stopped, so
- *  its builder's Message is `never` and no handler is expressible. Reload or
- *  navigate with a raw DOM attribute instead. */
-export type CrashConfig<Model, Message> = Readonly<{
-  view?: (
-    context: CrashContext<Model, Message>,
-    h: HtmlBuilder<never>,
-  ) => Document
-  report?: (context: CrashContext<Model, Message>) => void
-}>
-
 /** Full runtime configuration including Model Schema, Flags, init, update, view, and optional routing/stream config. */
-// NOTE: the payload scripts are held as elements, not as their text. Reading
-// the text is the first step of consuming another deployment's handoff, and the
-// build id has not been compared yet at the point this runs.
-type HydrationConfig = Readonly<{
-  root: HTMLElement
-  runtimeId: string
-  flagsScripts: ReadonlyArray<HTMLScriptElement>
-  isFlagsRequired: boolean
-}>
-
-const hydrationForRoot = (
-  root: HTMLElement,
-  isFlagsRequired: boolean,
-): HydrationConfig => {
-  const runtimeId = root.getAttribute(FOLDKIT_APP_ATTRIBUTE) ?? ''
-  const flagsScripts = pipe(
-    Array.fromIterable(
-      document.querySelectorAll<HTMLScriptElement>(
-        `script[${FOLDKIT_FLAGS_ATTRIBUTE}]`,
-      ),
-    ),
-    Array.filter(
-      script => script.getAttribute(FOLDKIT_FLAGS_ATTRIBUTE) === runtimeId,
-    ),
-  )
-  return { root, runtimeId, flagsScripts, isFlagsRequired }
-}
-
-// NOTE: hydration is scoped to the app's own stamped root so a server-rendered
-// app never adopts another app's DOM. A container that carries the stamp is that
-// root. A container that does not is a non-root element the caller resolved,
-// which happens when the rendered view has its own element with the container's
-// id (a descendant `id="root"`) and `getElementById` returned that inner element
-// instead of the stamped root above it. That inner element resolves to the app
-// root only when the page has exactly one stamped root and this container sits
-// inside it. A container outside that root is a refused handoff, not a fresh
-// application beside server markup that remains live. A null container is the
-// replace-parity case, where the server root took the placeholder's place and
-// `getElementById` no longer finds it, so the stamp is the only handle.
-// A runtime id names one application for the whole page: it pairs a root with
-// its Flags payload, and it keys the Model and scroll position hot reloading
-// preserves. Two roots sharing one are not two applications but one claimed
-// twice, so whichever boots second would read the other's handoff and restore
-// the other's Model. `injectIntoTemplate` refuses to build such a page; this is
-// the check for a page assembled some other way.
-//
-// Distinct ids do not make two hydrated applications independent, which is not
-// a supported arrangement: each page-owning application rewrites the document's
-// metadata and installs document-wide navigation listeners.
-const assertRuntimeIdsAreUnique = (
-  stampedRoots: ReadonlyArray<HTMLElement>,
-): void => {
-  const seen = new Set<string>()
-  for (const root of stampedRoots) {
-    const runtimeId = root.getAttribute(FOLDKIT_APP_ATTRIBUTE) ?? ''
-    if (runtimeId === '') {
-      containRefusedPage(root.ownerDocument)
-      throw new Error(
-        '[foldkit] Found a server-rendered root with an empty ' +
-          `\`${FOLDKIT_APP_ATTRIBUTE}\` stamp. A hydratable root must carry ` +
-          'a nonempty runtime id so the runtime can pair it with its Flags ' +
-          'payload and preserved HMR state.',
-      )
-    }
-    if (!root.ownerDocument.body?.contains(root)) {
-      containRefusedPage(root.ownerDocument)
-      throw new Error(
-        '[foldkit] Found a server-rendered root outside the document body. ' +
-          'Runtime.hydrate supports one page-owning application in the ' +
-          'document light DOM. Do not hydrate a root in the head, a shadow ' +
-          'tree, or a detached subtree.',
-      )
-    }
-    if (seen.has(runtimeId)) {
-      containRefusedPage(root.ownerDocument)
-      throw new Error(
-        `[foldkit] Found more than one server-rendered root stamped ` +
-          `"${runtimeId}". A runtime id names one application for the whole ` +
-          'page: it pairs a root with its Flags payload and keys the Model and ' +
-          'scroll position hot reloading preserves, so two roots sharing one ' +
-          "would take each other's state. Remove the duplicate root. Foldkit " +
-          'hydrates one page-owning application per document.',
-      )
-    }
-    seen.add(runtimeId)
-  }
-}
-
-const assertSinglePageApplication = (
-  stampedRoots: ReadonlyArray<HTMLElement>,
-): void => {
-  if (stampedRoots.length > 1) {
-    containRefusedPage(document)
-    throw new Error(
-      '[foldkit] Found more than one page-owning application stamped with ' +
-        `\`${FOLDKIT_APP_ATTRIBUTE}\`. Hydrating multiple applications in ` +
-        'one document is not supported: each application owns the document ' +
-        'metadata and installs document-wide navigation listeners. Render ' +
-        'one application per page.',
-    )
-  }
-}
-
-// The reason this page cannot be adopted by this client, or `undefined` when
-// the two name the same deployment.
-//
-// The client's id is required and must be non-empty. An absent one would
-// otherwise equal the absent marker on a page served before build ids existed,
-// which reads a page from an unknown deployment as one of this build's own: the
-// exact case the id exists to refuse.
-const buildSkew = (
-  root: HTMLElement,
-  buildId: unknown,
-  runtimeId: string,
-): Error | undefined => {
-  if (!Predicate.isString(buildId) || buildId === '') {
-    return new Error(
-      '[foldkit] Runtime.hydrate was given no build id. Hydration compares ' +
-        'the id the server stamped on the root with this client’s own before ' +
-        'it adopts any DOM, and without one a page from any deployment would ' +
-        'be adopted as this one. Pass ' +
-        '`buildId: import.meta.env.FOLDKIT_BUILD_ID`, the same value the ' +
-        'server entry passes to `renderToString`.',
-    )
-  }
-  const servedBuild = root.getAttribute(HYDRATION_BUILD_ATTRIBUTE)
-  if (servedBuild === buildId) {
-    return undefined
-  }
-  return new Error(
-    `[foldkit] Runtime.hydrate found application "${runtimeId}" served by ` +
-      `${servedBuild === null ? 'no known deployment' : `deployment "${servedBuild}"`}` +
-      `, but this client belongs to deployment "${buildId}". Startup stops ` +
-      'here rather than reading a handoff written by other code: the Flags in ' +
-      'the page are that deployment’s, and this build could accept them while ' +
-      'every value in them means something else. Serve the page from the ' +
-      'running deployment, and keep stale HTML out of shared caches.',
-  )
-}
-
-// Take the served page out of reach before startup stops.
-//
-// Refusing to adopt a page keeps this build's code away from it, but the markup
-// is still a live document: its links navigate, its forms submit to whatever
-// action the deployment that wrote them intended, and its controls take focus.
-// A visitor who clicks after a failed handoff would be acting on a page with no
-// running code to reconsider.
-//
-// The boundary is an element the served page already has, marked in place.
-// Nothing moves. Wrapping the root in a fresh inert element would reparent it:
-// every upgraded custom element in the subtree would run `disconnectedCallback`
-// and then `connectedCallback` again, and every embedded browsing context would
-// reload. Marking a stable ancestor costs neither.
-//
-// That ancestor is the document's body, which is where a hydratable root always
-// sits: `renderToString` refuses to render `html`, `head`, or `body` as the
-// root, and `hydrate` is for an application that owns the page. Inertness
-// propagates from an inert HTML element to every descendant whatever their
-// namespace, so an SVG or MathML root is covered too. `inert` is an HTML
-// attribute, so marking such a root directly would only create an expando.
-// Verified in Chromium for HTML, SVG foreign content, and MathML.
-//
-// A modal dialog lives in the top layer, where ancestor inertness does not reach
-// it. The refusal shield is itself a modal dialog opened after the served page's
-// top-layer content, so it covers an old dialog even inside a closed shadow
-// root. It is a sibling of `body`, outside the inert and `aria-hidden` boundary,
-// covers the viewport, and refuses its cancel action. Existing dialogs are not
-// closed: calling author-owned `close` or `cancel` listeners while startup is
-// failing can run arbitrary stale code. Same-document input guards stop
-// physical keyboard input if older top-layer content requests focus.
-//
-// This runs when the client reaches the failure, so it cannot undo what the page
-// already did: subresources the parser fetched, custom elements that upgraded,
-// scripts the served deployment authored, or anything a visitor managed before
-// the client entry ran. Nor is this a script sandbox. A capture listener on
-// `window` or `document` runs before an event reaches the shield, and a timer or
-// listener can open newer top-layer UI. The boundary stops pointer and physical
-// keyboard input from activating the old page's same-document native links,
-// forms, and controls without reconnecting its DOM.
-// NOTE: this module declares its own `Document` (the page metadata a view
-// describes), so the DOM interface is reached through the global binding.
-type DomDocument = typeof document
-
-const REFUSED_ATTRIBUTE = 'data-foldkit-refused'
-const REFUSAL_SHIELD_ATTRIBUTE = 'data-foldkit-refusal-shield'
-const refusalShields = new WeakMap<DomDocument, HTMLDialogElement>()
-const REFUSAL_SHIELD_INPUT_EVENTS: ReadonlyArray<keyof HTMLElementEventMap> = [
-  'auxclick',
-  'click',
-  'contextmenu',
-  'dblclick',
-  'keydown',
-  'keypress',
-  'keyup',
-  'mousedown',
-  'mouseup',
-  'pointerdown',
-  'pointerup',
-  'touchend',
-  'touchstart',
-]
-const REFUSAL_DOCUMENT_INPUT_EVENTS: ReadonlyArray<keyof HTMLElementEventMap> =
-  ['keydown', 'keypress', 'keyup']
-
-const preventRefusalShieldInteraction = (event: Event): void => {
-  event.preventDefault()
-  event.stopImmediatePropagation()
-}
-
-// NOTE: `showModal` throws when the owning document is not fully active. A
-// foreign or detached document still needs the original Foldkit refusal rather
-// than a browser exception from its containment UI.
-const tryOpenRefusalShieldAsModal = (shield: HTMLDialogElement): boolean => {
-  if (typeof shield.showModal !== 'function') {
-    return false
-  }
-  try {
-    shield.showModal()
-    return true
-  } catch {
-    return false
-  }
-}
-
-const openRefusalShield = (shield: HTMLDialogElement): void => {
-  shield.inert = true
-  shield.setAttribute('inert', '')
-  if (shield.open && typeof shield.close === 'function') {
-    shield.close()
-  }
-  if (!tryOpenRefusalShieldAsModal(shield)) {
-    shield.setAttribute('open', '')
-  }
-  shield.inert = false
-  shield.removeAttribute('inert')
-  shield.focus({ preventScroll: true })
-}
-
-const installRefusalShield = (ownerDocument: DomDocument): void => {
-  const existing = refusalShields.get(ownerDocument)
-  if (existing !== undefined && existing.isConnected) {
-    openRefusalShield(existing)
-    return
-  }
-
-  const shield = ownerDocument.createElement('dialog')
-  shield.setAttribute(REFUSAL_SHIELD_ATTRIBUTE, '')
-  shield.setAttribute('aria-label', 'Page unavailable')
-  shield.setAttribute('aria-modal', 'true')
-  shield.setAttribute('closedby', 'none')
-  shield.tabIndex = -1
-  shield.textContent =
-    'This page could not start safely. Reload to get the current version.'
-  shield.style.alignItems = 'center'
-  shield.style.background = 'rgba(15, 23, 42, 0.96)'
-  shield.style.border = '0'
-  shield.style.boxSizing = 'border-box'
-  shield.style.color = 'white'
-  shield.style.font = '600 1rem/1.5 system-ui, sans-serif'
-  shield.style.display = 'grid'
-  shield.style.height = '100vh'
-  shield.style.inset = '0'
-  shield.style.margin = '0'
-  shield.style.maxHeight = 'none'
-  shield.style.maxWidth = 'none'
-  shield.style.overflow = 'hidden'
-  shield.style.padding = '2rem'
-  shield.style.position = 'fixed'
-  shield.style.touchAction = 'none'
-  shield.style.userSelect = 'none'
-  shield.style.width = '100vw'
-  shield.addEventListener('cancel', preventRefusalShieldInteraction)
-  for (const eventName of REFUSAL_SHIELD_INPUT_EVENTS) {
-    shield.addEventListener(eventName, preventRefusalShieldInteraction, {
-      capture: true,
-      passive: false,
-    })
-  }
-  for (const eventName of REFUSAL_DOCUMENT_INPUT_EVENTS) {
-    ownerDocument.addEventListener(eventName, preventRefusalShieldInteraction, {
-      capture: true,
-      passive: false,
-    })
-  }
-  ownerDocument.documentElement.appendChild(shield)
-  refusalShields.set(ownerDocument, shield)
-  openRefusalShield(shield)
-}
-
-const containRefusedPage = (ownerDocument: DomDocument): void => {
-  const boundary = ownerDocument.body ?? ownerDocument.documentElement
-  boundary.inert = true
-  boundary.setAttribute('inert', '')
-  boundary.setAttribute('aria-hidden', 'true')
-  boundary.setAttribute(REFUSED_ATTRIBUTE, '')
-  installRefusalShield(ownerDocument)
-}
-
-// Whether this page carries anything a Foldkit server render leaves behind. A
-// resolution failure on such a page is a refused handoff, and the markup is
-// contained; the same failure on a page with none of these markers is a client
-// application whose container never existed, where there is no server render to
-// refuse and nothing to take out of reach.
-const hasServerRenderedMarkup = (ownerDocument: DomDocument): boolean =>
-  ownerDocument.querySelector(
-    `[${FOLDKIT_APP_ATTRIBUTE}], [${HYDRATION_BUILD_ATTRIBUTE}], ` +
-      `[${FOLDKIT_FLAGS_ATTRIBUTE}]`,
-  ) !== null
-
-const findDocumentHydration = (
-  container: HTMLElement | null,
-  isFlagsRequired: boolean,
-): HydrationConfig | undefined => {
-  const stampedRoots = Array.fromIterable(
-    document.querySelectorAll<HTMLElement>(`[${FOLDKIT_APP_ATTRIBUTE}]`),
-  )
-  assertRuntimeIdsAreUnique(stampedRoots)
-  assertSinglePageApplication(stampedRoots)
-  if (container !== null) {
-    const stampedAncestor = container.closest<HTMLElement>(
-      `[${FOLDKIT_APP_ATTRIBUTE}]`,
-    )
-    if (container.hasAttribute(FOLDKIT_APP_ATTRIBUTE)) {
-      const isDocumentRoot = Option.match(Array.head(stampedRoots), {
-        onNone: () => false,
-        onSome: root => root === container,
-      })
-      if (!isDocumentRoot) {
-        containRefusedPage(container.ownerDocument)
-        throw new Error(
-          '[foldkit] Runtime.hydrate received a stamped container that is ' +
-            'not the single server root in the document light DOM. Hydration ' +
-            'supports one page-owning application under the document body. ' +
-            'Do not hydrate a root in a shadow tree or detached subtree.',
-        )
-      }
-      return hydrationForRoot(container, isFlagsRequired)
-    }
-    return Array.match(stampedRoots, {
-      onEmpty: () => {
-        if (stampedAncestor === null) {
-          return undefined
-        }
-        containRefusedPage(container.ownerDocument)
-        throw new Error(
-          '[foldkit] Runtime.hydrate received a container under a stamped ' +
-            'root outside the document body light DOM. Do not hydrate a root ' +
-            'in a shadow tree, detached subtree, or another document.',
-        )
-      },
-      onNonEmpty: roots => {
-        const onlyRoot = Array.headNonEmpty(roots)
-        if (stampedAncestor === onlyRoot) {
-          return hydrationForRoot(onlyRoot, isFlagsRequired)
-        }
-        containRefusedPage(container.ownerDocument)
-        throw new Error(
-          '[foldkit] Runtime.hydrate received a container outside the ' +
-            "document's server-rendered application root. A page-owning " +
-            'application must adopt that single root rather than boot beside ' +
-            'server markup it does not own.',
-        )
-      },
-    })
-  }
-  return Option.match(Array.head(stampedRoots), {
-    onNone: () => undefined,
-    onSome: root => hydrationForRoot(root, isFlagsRequired),
-  })
-}
-
 type RuntimeConfig<
   Model,
   Message,
@@ -1265,14 +432,6 @@ export type ApplicationConfig<
     >
   }>
 
-/** Configuration for crash handling in a `makeElement` app. The crash view
- *  returns `Html`, not a `Document`, because a scoped app never owns the
- *  document `<head>`. */
-export type ElementCrashConfig<Model, Message> = Readonly<{
-  view?: (context: CrashContext<Model, Message>, h: HtmlBuilder<never>) => Html
-  report?: (context: CrashContext<Model, Message>) => void
-}>
-
 type BaseElementConfig<
   Model,
   Message,
@@ -1417,268 +576,6 @@ export type MakeRuntimeReturn<
   }>
 }>
 
-/** Host-side handle for one inbound Port. `send` validates the value by
- *  decoding it against the Port's Schema: on success the decoded value enters
- *  the app through the Port's Subscription; on failure nothing reaches the
- *  app, the failure is logged, and the returned `Exit` carries the
- *  `SchemaError`. Sends after `dispose` are no-ops. */
-export type InboundPortHandle<Encoded> = Readonly<{
-  send: (value: Encoded) => Exit.Exit<void, Schema.SchemaError>
-}>
-
-/** Host-side handle for one outbound Port. `subscribe` registers a listener
- *  for the encoded values the app emits with `Port.emit` and returns an
- *  unsubscribe function. Multiple listeners receive each value in
- *  registration order. */
-export type OutboundPortHandle<Encoded> = Readonly<{
-  subscribe: (listener: (value: Encoded) => void) => () => void
-}>
-
-/** The inbound half of `PortHandles`: one `InboundPortHandle` per declared
- *  inbound Port, keyed by Port name. */
-export type InboundPortHandles<InboundPorts> =
-  InboundPorts extends Readonly<Record<string, Inbound<any, any>>>
-    ? {
-        readonly [
-          Name in keyof InboundPorts
-        ]: InboundPorts[Name] extends Inbound<any, infer Encoded>
-          ? InboundPortHandle<Encoded>
-          : never
-      }
-    : unknown
-
-/** The outbound half of `PortHandles`: one `OutboundPortHandle` per declared
- *  outbound Port, keyed by Port name. */
-export type OutboundPortHandles<OutboundPorts> =
-  OutboundPorts extends Readonly<Record<string, Outbound<any, any>>>
-    ? {
-        readonly [
-          Name in keyof OutboundPorts
-        ]: OutboundPorts[Name] extends Outbound<any, infer Encoded>
-          ? OutboundPortHandle<Encoded>
-          : never
-      }
-    : unknown
-
-/** The `ports` field of an `EmbedHandle`: one `InboundPortHandle` or
- *  `OutboundPortHandle` per declared Port, keyed by Port name. */
-export type PortHandles<P extends Ports | undefined> = P extends Ports
-  ? InboundPortHandles<P['inbound']> & OutboundPortHandles<P['outbound']>
-  : unknown
-
-/**
- * The handle returned by `embed`. The host talks to the embedded app only
- * through it: `ports.<name>.send` pushes values in, `ports.<name>.subscribe`
- * listens to values the app emits, and `dispose` shuts the runtime down.
- *
- * `dispose` is idempotent. It interrupts the runtime and runs all cleanup:
- * Subscriptions, ManagedResources, Mounts, listeners, and in-flight Commands
- * stop, and the rendered DOM is removed with the container element restored
- * empty in its place, ready for a fresh `embed`.
- */
-export type EmbedHandle<P extends Ports | undefined = undefined> = Readonly<{
-  ports: PortHandles<P>
-  dispose: () => void
-}>
-
-type HostConnector = Readonly<{
-  sendInbound: (
-    portName: string,
-    port: Inbound<any, any>,
-    value: unknown,
-  ) => Exit.Exit<void, Schema.SchemaError>
-  addListener: (
-    port: Outbound<any, any>,
-    listener: (encodedValue: unknown) => void,
-  ) => () => void
-  deliverOutbound: (port: Outbound<any, any>, encodedValue: unknown) => void
-  bind: (
-    deliverInbound: (port: Inbound<any, any>, value: unknown) => void,
-  ) => void
-  unbind: () => void
-  dispose: () => void
-}>
-
-const makeHostConnector = (): HostConnector => {
-  let isDisposed = false
-  let maybeDeliverInbound: Option.Option<
-    (port: Inbound<any, any>, value: unknown) => void
-  > = Option.none()
-  const pendingInboundSends: Array<{
-    port: Inbound<any, any>
-    value: unknown
-  }> = []
-  const listenersByPort = new Map<
-    Outbound<any, any>,
-    Set<(encodedValue: unknown) => void>
-  >()
-
-  const sendInbound = (
-    portName: string,
-    port: Inbound<any, any>,
-    value: unknown,
-  ): Exit.Exit<void, Schema.SchemaError> => {
-    if (isDisposed) {
-      return Exit.void
-    }
-    const decodeExit = Schema.decodeUnknownExit(port.schema)(value)
-    Exit.match(decodeExit, {
-      onFailure: cause => {
-        console.error(
-          `[foldkit] Inbound port "${portName}" rejected a value:`,
-          Cause.squash(cause),
-        )
-      },
-      onSuccess: decodedValue => {
-        Option.match(maybeDeliverInbound, {
-          onNone: () => {
-            pendingInboundSends.push({ port, value: decodedValue })
-          },
-          onSome: deliverInbound => deliverInbound(port, decodedValue),
-        })
-      },
-    })
-    return Exit.asVoid(decodeExit)
-  }
-
-  const addListener = (
-    port: Outbound<any, any>,
-    listener: (encodedValue: unknown) => void,
-  ): (() => void) => {
-    if (isDisposed) {
-      return Function.constVoid
-    }
-    const listeners = listenersByPort.get(port) ?? new Set()
-    listenersByPort.set(port, listeners)
-    listeners.add(listener)
-    return () => {
-      listeners.delete(listener)
-    }
-  }
-
-  // NOTE: delivery is deferred to a microtask so a host listener never runs
-  // inside the runtime's Command fiber (a listener that synchronously calls
-  // send or dispose must not re-enter the runtime), and so a host that
-  // subscribes synchronously right after embed() returns still receives
-  // emissions from init Commands.
-  const deliverOutbound = (
-    port: Outbound<any, any>,
-    encodedValue: unknown,
-  ): void => {
-    if (isDisposed) {
-      return
-    }
-    queueMicrotask(() => {
-      if (isDisposed) {
-        return
-      }
-      const listeners = listenersByPort.get(port) ?? new Set()
-      listeners.forEach(listener => {
-        try {
-          listener(encodedValue)
-        } catch (listenerError) {
-          console.error(
-            '[foldkit] An outbound port listener threw:',
-            listenerError,
-          )
-        }
-      })
-    })
-  }
-
-  const bind = (
-    deliverInbound: (port: Inbound<any, any>, value: unknown) => void,
-  ): void => {
-    maybeDeliverInbound = Option.some(deliverInbound)
-    const flushedSends = pendingInboundSends.splice(0)
-    flushedSends.forEach(({ port, value }) => deliverInbound(port, value))
-  }
-
-  const unbind = (): void => {
-    maybeDeliverInbound = Option.none()
-  }
-
-  const dispose = (): void => {
-    isDisposed = true
-    pendingInboundSends.length = 0
-    listenersByPort.forEach(listeners => listeners.clear())
-    listenersByPort.clear()
-  }
-
-  return { sendInbound, addListener, deliverOutbound, bind, unbind, dispose }
-}
-
-type PortChannelsBundle = Readonly<{
-  channels: __PortChannels
-  deliverInbound: (port: Inbound<any, any>, value: unknown) => void
-}>
-
-const makePortChannels = (
-  ports: Ports,
-  maybeConnector: Option.Option<HostConnector>,
-): PortChannelsBundle => {
-  const inboundChannelsByPort = new Map<Inbound<any, any>, __InboundChannel>()
-  Object.values(ports.inbound ?? {}).forEach(port => {
-    inboundChannelsByPort.set(port, __makeInboundChannel())
-  })
-
-  const outboundPorts = new Set(Object.values(ports.outbound ?? {}))
-
-  const channels: __PortChannels = {
-    isConfigured: true,
-    lookupInbound: port =>
-      Option.fromNullishOr(inboundChannelsByPort.get(port)),
-    lookupOutbound: port =>
-      outboundPorts.has(port)
-        ? Option.some(encodedValue =>
-            Option.match(maybeConnector, {
-              onNone: Function.constVoid,
-              onSome: connector =>
-                connector.deliverOutbound(port, encodedValue),
-            }),
-          )
-        : Option.none(),
-  }
-
-  const deliverInbound = (port: Inbound<any, any>, value: unknown): void => {
-    Option.match(Option.fromNullishOr(inboundChannelsByPort.get(port)), {
-      onNone: Function.constVoid,
-      onSome: channel => channel.deliver(value),
-    })
-  }
-
-  return { channels, deliverInbound }
-}
-
-const validatePorts = (ports: Ports): void => {
-  const inboundEntries = Object.entries(ports.inbound ?? {})
-  const outboundEntries = Object.entries(ports.outbound ?? {})
-
-  const inboundNames = new Set(inboundEntries.map(([name]) => name))
-  outboundEntries.forEach(([name]) => {
-    if (inboundNames.has(name)) {
-      throw new Error(
-        `[foldkit] Port name "${name}" appears in both inbound and outbound. ` +
-          'Port names share one namespace on the EmbedHandle, so each name ' +
-          'must be unique across both records.',
-      )
-    }
-  })
-
-  const seenPorts = new Set<unknown>()
-  const allEntries = [...inboundEntries, ...outboundEntries]
-  allEntries.forEach(([name, port]) => {
-    if (seenPorts.has(port)) {
-      throw new Error(
-        `[foldkit] The Port registered as "${name}" is also registered under ` +
-          'another name. Each entry in the ports record needs its own ' +
-          'Port.inbound or Port.outbound value.',
-      )
-    }
-    seenPorts.add(port)
-  })
-}
-
 type RuntimeInternals = {
   startWith: (
     maybeConnector: Option.Option<HostConnector>,
@@ -1741,11 +638,7 @@ const makeRuntime = <
   Kind
 >): MakeRuntimeReturn<P, Flags, Resources, Kind> => {
   const isSlowVisible = (show: Visibility): boolean =>
-    Match.value(show).pipe(
-      Match.when('Always', () => true),
-      Match.when('Development', () => !!import.meta.hot),
-      Match.exhaustive,
-    )
+    isVisible(show, !!import.meta.hot)
 
   const htmlBuilder = htmlBuilderFor<Message>()
 
@@ -1790,40 +683,9 @@ const makeRuntime = <
     ? createDuplicateIdScanner()
     : undefined
 
-  const excludeFromHistoryTags: ReadonlySet<string> = pipe(
-    devTools ?? {},
-    Option.liftPredicate(config => config !== false),
-    Option.flatMapNullishOr(config => config.excludeFromHistory),
-    Option.match({
-      onNone: () => new Set<string>(),
-      onSome: tags => new Set(tags),
-    }),
-  )
-
-  const devToolsMaxEntries: number | undefined = pipe(
-    devTools ?? {},
-    Option.liftPredicate(config => config !== false),
-    Option.flatMapNullishOr(config => config.maxEntries),
-    Option.match({
-      onNone: () => undefined,
-      onSome: value =>
-        Math.max(
-          DEV_TOOLS_MAX_ENTRIES_MIN,
-          Math.min(DEV_TOOLS_MAX_ENTRIES_MAX, value),
-        ),
-    }),
-  )
-
-  const devToolsKeyframeInterval: number | undefined = pipe(
-    devTools ?? {},
-    Option.liftPredicate(config => config !== false),
-    Option.flatMapNullishOr(config => config.keyframeInterval),
-    Option.match({
-      onNone: () => undefined,
-      onSome: value =>
-        Math.max(DEV_TOOLS_KEYFRAME_INTERVAL_MIN, Math.floor(value)),
-    }),
-  )
+  const excludeFromHistoryTags = resolveExcludeFromHistoryTags(devTools)
+  const devToolsMaxEntries = resolveDevToolsMaxEntries(devTools)
+  const devToolsKeyframeInterval = resolveDevToolsKeyframeInterval(devTools)
 
   const maybeFreezeModel = (model: Model): Model =>
     isFreezeModelActive ? deepFreeze(model) : model
@@ -2394,24 +1256,7 @@ const makeRuntime = <
 
         const initModel = maybeFreezeModel(initModelRaw)
 
-        const isInIframe = window.self !== window.top
-        const resolvedDevTools = pipe(
-          devTools ?? {},
-          Option.liftPredicate(config => config !== false),
-          Option.filter(config =>
-            Match.value(config.show ?? DEFAULT_DEV_TOOLS_SHOW).pipe(
-              Match.when('Always', () => true),
-              Match.when('Development', () => !!import.meta.hot && !isInIframe),
-              Match.exhaustive,
-            ),
-          ),
-          Option.map(config => ({
-            position: config.position ?? DEFAULT_DEV_TOOLS_POSITION,
-            mode: resolveDevToolsMode(config.mode ?? DEFAULT_DEV_TOOLS_MODE),
-            maybeBanner: Option.fromNullishOr(config.banner),
-            maybeOverlay: Option.fromNullishOr(registeredDevToolsOverlay),
-          })),
-        )
+        const resolvedDevTools = resolveDevToolsConfig(devTools)
 
         const modelPubSub = yield* PubSub.unbounded<Model>()
         let currentViewState: ViewState = 'Live'
@@ -3759,315 +2604,6 @@ const makeRuntime = <
 // reads the clock beyond the drain start.
 const DRAIN_BUDGET_MS = 5
 
-/** Mutable holder for the vnode tree currently mounted in the container.
- *  The render frame writes it after every patch; the dispose finalizer, the
- *  replay render, and {@link renderCrashView} read it. A plain object rather
- *  than a Ref because every reader runs synchronously on the main thread. */
-type VNodeSlot = {
-  maybeCurrentVNode: Option.Option<VNode>
-}
-
-const currentLocationUrl = (): string => {
-  const { origin, pathname, search } = window.location
-  return `${origin}${pathname}${search}`
-}
-
-type DocumentMetadataElements = {
-  canonical?: HTMLLinkElement
-  ogUrl?: HTMLMetaElement
-}
-
-type ResolvedDocumentMetadata = Readonly<{
-  title: string
-  lang: string | undefined
-  dirAttribute: string | undefined
-  canonicalUrl: string
-  ogUrl: string
-}>
-
-type DocumentMetadataInvalidation = { isInvalidated: boolean }
-
-type DocumentMetadataState = {
-  readonly elements: DocumentMetadataElements
-  readonly observer: MutationObserver
-  readonly invalidation: DocumentMetadataInvalidation
-  observedHead: HTMLHeadElement
-  lastApplied?: ResolvedDocumentMetadata
-  cachedLocationHref?: string
-  cachedLocationCanonical?: string
-}
-
-const documentMetadataStates = new WeakMap<
-  globalThis.Document,
-  DocumentMetadataState
->()
-
-const observeDocumentMetadataMutations = (
-  observer: MutationObserver,
-): HTMLHeadElement => {
-  const observedHead = document.head
-  observer.observe(observedHead, {
-    attributes: true,
-    attributeFilter: ['rel', 'href', 'property', 'content'],
-    childList: true,
-    characterData: true,
-    subtree: true,
-  })
-  observer.observe(document.documentElement, {
-    attributes: true,
-    attributeFilter: ['lang', 'dir'],
-  })
-  return observedHead
-}
-
-const getOrCreateDocumentMetadataState = (): DocumentMetadataState => {
-  const existingState = documentMetadataStates.get(document)
-  if (existingState !== undefined) {
-    return existingState
-  }
-
-  const invalidation: DocumentMetadataInvalidation = { isInvalidated: true }
-  const observer = new MutationObserver(() => {
-    invalidation.isInvalidated = true
-  })
-  const metadataState: DocumentMetadataState = {
-    elements: {},
-    observer,
-    invalidation,
-    observedHead: observeDocumentMetadataMutations(observer),
-  }
-  documentMetadataStates.set(document, metadataState)
-  return metadataState
-}
-
-const findOrCreateDocumentMetadataElements = (
-  metadataState: DocumentMetadataState,
-): Readonly<{
-  canonical: HTMLLinkElement
-  ogUrl: HTMLMetaElement
-}> => {
-  const { elements } = metadataState
-
-  let canonical = elements.canonical
-  if (canonical === undefined || canonical.parentNode !== document.head) {
-    canonical =
-      document.head.querySelector<HTMLLinkElement>('link[rel="canonical"]') ??
-      document.head.appendChild(document.createElement('link'))
-    elements.canonical = canonical
-  }
-
-  let ogUrl = elements.ogUrl
-  if (ogUrl === undefined || ogUrl.parentNode !== document.head) {
-    ogUrl =
-      document.head.querySelector<HTMLMetaElement>('meta[property="og:url"]') ??
-      document.head.appendChild(document.createElement('meta'))
-    elements.ogUrl = ogUrl
-  }
-
-  return { canonical, ogUrl }
-}
-
-const readOrCacheCurrentLocationUrl = (
-  metadataState: DocumentMetadataState,
-): string => {
-  const currentLocationHref = window.location.href
-  if (
-    metadataState.cachedLocationHref === currentLocationHref &&
-    metadataState.cachedLocationCanonical !== undefined
-  ) {
-    return metadataState.cachedLocationCanonical
-  }
-
-  const canonicalUrl = currentLocationUrl()
-  metadataState.cachedLocationHref = currentLocationHref
-  metadataState.cachedLocationCanonical = canonicalUrl
-  return canonicalUrl
-}
-
-const rebindDocumentMetadataObserverToCurrentHead = (
-  metadataState: DocumentMetadataState,
-): void => {
-  // NOTE: a MutationObserver remains attached to a detached head after
-  // document.head is replaced. Rebind it and invalidate the metadata snapshot
-  // before the next unchanged guard.
-  metadataState.observer.disconnect()
-  metadataState.observedHead = observeDocumentMetadataMutations(
-    metadataState.observer,
-  )
-  metadataState.invalidation.isInvalidated = true
-}
-
-const reconcileDocumentMetadata = (
-  metadataState: DocumentMetadataState,
-  nextMetadata: ResolvedDocumentMetadata,
-): void => {
-  if (document.title !== nextMetadata.title) {
-    document.title = nextMetadata.title
-  }
-
-  const { documentElement } = document
-
-  if (
-    nextMetadata.lang !== undefined &&
-    documentElement.lang !== nextMetadata.lang
-  ) {
-    documentElement.lang = nextMetadata.lang
-  }
-
-  if (
-    nextMetadata.dirAttribute !== undefined &&
-    documentElement.dir !== nextMetadata.dirAttribute
-  ) {
-    documentElement.dir = nextMetadata.dirAttribute
-  }
-
-  const metadataElements = findOrCreateDocumentMetadataElements(metadataState)
-
-  if (metadataElements.canonical.getAttribute('rel') !== 'canonical') {
-    metadataElements.canonical.setAttribute('rel', 'canonical')
-  }
-  if (
-    metadataElements.canonical.getAttribute('href') !==
-    nextMetadata.canonicalUrl
-  ) {
-    metadataElements.canonical.setAttribute('href', nextMetadata.canonicalUrl)
-  }
-  if (metadataElements.ogUrl.getAttribute('property') !== 'og:url') {
-    metadataElements.ogUrl.setAttribute('property', 'og:url')
-  }
-  if (metadataElements.ogUrl.getAttribute('content') !== nextMetadata.ogUrl) {
-    metadataElements.ogUrl.setAttribute('content', nextMetadata.ogUrl)
-  }
-
-  metadataState.lastApplied = nextMetadata
-  metadataState.invalidation.isInvalidated = false
-  // NOTE: clear the records produced by Foldkit's own writes before the
-  // observer callback runs. Otherwise the next unchanged render would be
-  // marked dirty and repeat every DOM read this cache is meant to avoid.
-  metadataState.observer.takeRecords()
-}
-
-const applyDocumentMetadata = (
-  nextDocument: Document,
-  mountedRoot: Node | undefined,
-): void => {
-  if (!mountedRoot || !document.body.contains(mountedRoot)) {
-    return
-  }
-
-  const metadataState = getOrCreateDocumentMetadataState()
-
-  if (metadataState.observedHead !== document.head) {
-    rebindDocumentMetadataObserverToCurrentHead(metadataState)
-  }
-
-  const canonicalUrl =
-    nextDocument.canonical ?? readOrCacheCurrentLocationUrl(metadataState)
-  const ogUrl = nextDocument.ogUrl ?? canonicalUrl
-  const dirAttribute =
-    nextDocument.dir === undefined
-      ? undefined
-      : textDirectionToAttribute(nextDocument.dir)
-
-  // NOTE: MutationObserver callbacks are asynchronous. Consume queued records
-  // synchronously before trusting the unchanged fast path.
-  if (Array.isArrayNonEmpty(metadataState.observer.takeRecords())) {
-    metadataState.invalidation.isInvalidated = true
-  }
-
-  const lastApplied = metadataState.lastApplied
-  const isAppliedMetadataCurrent =
-    !metadataState.invalidation.isInvalidated &&
-    lastApplied !== undefined &&
-    lastApplied.title === nextDocument.title &&
-    lastApplied.lang === nextDocument.lang &&
-    lastApplied.dirAttribute === dirAttribute &&
-    lastApplied.canonicalUrl === canonicalUrl &&
-    lastApplied.ogUrl === ogUrl
-
-  if (isAppliedMetadataCurrent) {
-    return
-  }
-
-  reconcileDocumentMetadata(metadataState, {
-    title: nextDocument.title,
-    lang: nextDocument.lang,
-    dirAttribute,
-    canonicalUrl,
-    ogUrl,
-  })
-}
-
-const renderCrashView = <Model, Message>(
-  context: CrashContext<Model, Message>,
-  crash: CrashConfig<Model, Message> | undefined,
-  container: HTMLElement,
-  vnodeSlot: VNodeSlot,
-  manageDocument: boolean,
-): void => {
-  console.error('[foldkit] Application crash:', context.error)
-
-  if (crash?.report) {
-    try {
-      crash.report(context)
-    } catch (reportError) {
-      console.error('[foldkit] crash.report failed:', reportError)
-    }
-  }
-
-  const crashContext = Context.make(Dispatch, noOpDispatch).pipe(
-    Context.add(MountTracker, {
-      started: () => {},
-      ended: () => {},
-    }),
-  )
-
-  try {
-    setHtmlRuntime(noOpDispatch.dispatchSync, crashContext)
-    let crashDocument: Document
-    try {
-      crashDocument = crash?.view
-        ? crash.view(context, htmlBuilderFor<never>())
-        : defaultCrashView(context)
-    } finally {
-      clearHtmlRuntime()
-    }
-
-    const patchedVNode = __patchVNode(
-      vnodeSlot.maybeCurrentVNode,
-      crashDocument.body,
-      container,
-    )
-    vnodeSlot.maybeCurrentVNode = Option.some(patchedVNode)
-    if (manageDocument) {
-      applyDocumentMetadata(crashDocument, patchedVNode.elm)
-    }
-  } catch (viewError) {
-    console.error('[foldkit] crash.view failed:', viewError)
-
-    const fallbackViewError =
-      viewError instanceof Error ? viewError : new Error(String(viewError))
-
-    setHtmlRuntime(noOpDispatch.dispatchSync, crashContext)
-    let fallbackDocument: Document
-    try {
-      fallbackDocument = defaultCrashView(context, fallbackViewError)
-    } finally {
-      clearHtmlRuntime()
-    }
-
-    const patchedVNode = __patchVNode(
-      vnodeSlot.maybeCurrentVNode,
-      fallbackDocument.body,
-      container,
-    )
-    vnodeSlot.maybeCurrentVNode = Option.some(patchedVNode)
-    if (manageDocument) {
-      applyDocumentMetadata(fallbackDocument, patchedVNode.elm)
-    }
-  }
-}
-
 /** Creates a Foldkit application that owns the page and returns a runtime that
  *  can be passed to `run`. The `view` returns a `Document`, so the runtime
  *  manages `document.title` and the canonical / og:url tags. Add a `routing`
@@ -4527,100 +3063,6 @@ export function makeElement<
   /* eslint-enable @typescript-eslint/consistent-type-assertions */
 }
 
-const encodePreserveModelMessage =
-  Schema.encodeUnknownSync(PreserveModelMessage)
-const encodeRequestModelMessage = Schema.encodeUnknownSync(RequestModelMessage)
-const decodeRestoreModelMessage = Schema.decodeUnknownExit(RestoreModelMessage)
-
-const preserveModel = (
-  id: string,
-  encodedModel: unknown,
-  isHmrReload: boolean,
-): void => {
-  if (import.meta.hot) {
-    import.meta.hot.send(
-      'foldkit:preserve-model',
-      encodePreserveModelMessage(
-        PreserveModelMessage.make({ id, model: encodedModel, isHmrReload }),
-      ),
-    )
-  }
-}
-
-const PLUGIN_RESPONSE_TIMEOUT_MS = 500
-
-// NOTE: scheduling fix for browser performance. Effect needs to defer work
-// onto a future tick of the event loop. The default browser scheduler picks
-// `setTimeout(f, 0)`, but browsers clamp `setTimeout` to a minimum of 4ms.
-// `queueMicrotask` runs on the very next tick (sub-millisecond). Dispatch no
-// longer routes through the Effect scheduler, but Command and Subscription
-// fibers still do; without this override every fiber yield (for example, an
-// op-budget suspension, or a Stream step) would take an extra 4-16ms
-// round-trip before
-// its result Message lands.
-const microtaskSetImmediate = (callback: () => void): (() => void) => {
-  let cancelled = false
-  queueMicrotask(() => {
-    if (!cancelled) callback()
-  })
-  return () => {
-    cancelled = true
-  }
-}
-
-const browserScheduler = new Scheduler.MixedScheduler(
-  'async',
-  microtaskSetImmediate,
-)
-
-const provideBrowserScheduler = <A, E, R>(
-  effect: Effect.Effect<A, E, R>,
-): Effect.Effect<A, E, R> =>
-  Effect.provide(effect, Layer.succeed(Scheduler.Scheduler, browserScheduler))
-
-// NOTE: asks @foldkit/vite-plugin for a model preserved across the last HMR
-// reload. The plugin only serves a model whose preservation was flushed by a
-// reload, so a host-driven dispose-then-embed remount initializes fresh while
-// a code reload restores state.
-const resolveHmrModel = (runtimeId: string): Effect.Effect<unknown> => {
-  const hot = import.meta.hot
-  if (!hot) {
-    return Effect.succeed(undefined)
-  }
-
-  return pipe(
-    Effect.callback<unknown>(resume => {
-      const handler = (message: unknown): void => {
-        Exit.match(decodeRestoreModelMessage(message), {
-          onFailure: Function.constVoid,
-          onSuccess: ({ id, model }) => {
-            if (id === runtimeId) {
-              hot.off('foldkit:restore-model', handler)
-              resume(Effect.succeed(model))
-            }
-          },
-        })
-      }
-      hot.on('foldkit:restore-model', handler)
-      hot.send(
-        'foldkit:request-model',
-        encodeRequestModelMessage(RequestModelMessage.make({ id: runtimeId })),
-      )
-      return Effect.sync(() => hot.off('foldkit:restore-model', handler))
-    }),
-    Effect.timeout(PLUGIN_RESPONSE_TIMEOUT_MS),
-    Effect.catchTag('TimeoutError', () => {
-      console.warn(
-        '[foldkit] No response from @foldkit/vite-plugin. Add it to your vite.config.ts for HMR model preservation:\n\n' +
-          "  import { foldkit } from '@foldkit/vite-plugin'\n\n" +
-          '  export default defineConfig({ plugins: [foldkit()] })\n\n' +
-          'Starting without HMR support.',
-      )
-      return Effect.succeed(undefined)
-    }),
-  )
-}
-
 /** Starts a program Effect with explicit boot inputs for runtime tests.
  * @internal */
 export const __startProgram = (
@@ -4750,30 +3192,6 @@ export const hydrate = <P extends Ports | undefined, Flags, Resources>(
   options: HydrateOptions,
 ): void => {
   startProgram(program, 'Hydrate', undefined, options?.buildId)
-}
-
-const buildPortHandles = <P extends Ports | undefined>(
-  ports: P,
-  connector: HostConnector,
-): PortHandles<P> => {
-  const handles: Record<string, unknown> = {}
-
-  if (Predicate.isNotUndefined(ports)) {
-    Object.entries(ports.inbound ?? {}).forEach(([portName, port]) => {
-      handles[portName] = {
-        send: (value: unknown) => connector.sendInbound(portName, port, value),
-      }
-    })
-    Object.entries(ports.outbound ?? {}).forEach(([portName, port]) => {
-      handles[portName] = {
-        subscribe: (listener: (encodedValue: unknown) => void) =>
-          connector.addListener(port, listener),
-      }
-    })
-  }
-
-  /* eslint-disable-next-line @typescript-eslint/consistent-type-assertions */
-  return handles as PortHandles<P>
 }
 
 /**
